+++ src/idunno.Bluesky/Actor/ActivitySubscriptions.cs
@@ -8,7 +8,7 @@ namespace idunno.Bluesky.Actor;
 /// <summary>
 /// A user's subscription preferences for other actors.
 /// </summary>
-/// <param name="allowSubscriptions">A <see cref="NotificationAllowedFrom"/> setting for the actor.</param>
-public sealed record ActivitySubscriptions(NotificationAllowedFrom allowSubscriptions)
+/// <param name="AllowSubscriptions">A <see cref="NotificationAllowedFrom"/> setting for the actor.</param>
+public sealed record ActivitySubscriptions(NotificationAllowedFrom AllowSubscriptions)
 {
 }