+++ src/idunno.Bluesky/Embed/Gallery/GalleryImage.cs
@@ -5,24 +5,24 @@ using System.Text.Json.Serialization;
 
 using idunno.AtProto;
 
-namespace idunno.Bluesky.Embed;
+namespace idunno.Bluesky.Embed.Gallery;
 
 /// <summary>
 /// Encapsulates information about an image embedded in a gallery.
 /// </summary>
 [JsonPolymorphic(IgnoreUnrecognizedTypeDiscriminators = false, UnknownDerivedTypeHandling = JsonUnknownDerivedTypeHandling.FailSerialization)]
-[JsonDerivedType(typeof(EmbeddedGalleryImage), typeDiscriminator: "app.bsky.embed.gallery#image")]
-public record EmbeddedGalleryImage
+[JsonDerivedType(typeof(GalleryImage), typeDiscriminator: "app.bsky.embed.gallery#image")]
+public record GalleryImage
 {
     /// <summary>
-    /// Creates a new instance of <see cref="EmbeddedGalleryImage"/>
+    /// Creates a new instance of <see cref="Image"/>
     /// </summary>
     /// <param name="image">A <see cref="Blob"/> containing details of the uploaded image.</param>
     /// <param name="altText">The AltText (for accessibility) for the image.</param>
     /// <param name="aspectRatio">The image's aspect ratio.</param>
     /// <exception cref="ArgumentNullException">Thrown when any of the parameters are <see langword="null" />.</exception>
     [JsonConstructor]
-    public EmbeddedGalleryImage(Blob image, string altText, AspectRatio aspectRatio)
+    public GalleryImage(Blob image, string altText, AspectRatio aspectRatio)
     {
         ArgumentNullException.ThrowIfNull(image);
         ArgumentNullException.ThrowIfNull(altText);
@@ -34,11 +34,11 @@ public record EmbeddedGalleryImage
     }
 
     /// <summary>
-    /// Creates a new instance of <see cref="EmbeddedGalleryImage"/> from the specified <paramref name="image"/>.
+    /// Creates a new instance of <see cref="Image"/> from the specified <paramref name="image"/>.
     /// </summary>
     /// <param name="image">The <see cref="EmbeddedImage"/> to create the gallery image from.</param>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="image"/> or any of its properties are <see langword="null"/>.</exception>
-    public EmbeddedGalleryImage(EmbeddedImage image)
+    public GalleryImage(EmbeddedImage image)
     {
         ArgumentNullException.ThrowIfNull(image);
         ArgumentNullException.ThrowIfNull(image.Image);