+++ src/idunno.Bluesky/Embed/View.cs
@@ -3,10 +3,6 @@
 
 using System.Text.Json.Serialization;
 
-using idunno.Bluesky.Feed;
-using idunno.Bluesky.Graph;
-using idunno.Bluesky.Labeler;
-
 namespace idunno.Bluesky.Embed;
 
 /// <summary>
@@ -17,10 +13,6 @@ namespace idunno.Bluesky.Embed;
 [JsonDerivedType(typeof(ViewNotFound), typeDiscriminator: EmbeddedViewTypeDiscriminators.EmbedViewNotFound)]
 [JsonDerivedType(typeof(ViewBlocked), typeDiscriminator: EmbeddedViewTypeDiscriminators.EmbedViewBlocked)]
 [JsonDerivedType(typeof(ViewDetached), typeDiscriminator: EmbeddedViewTypeDiscriminators.EmbedViewDetached)]
-[JsonDerivedType(typeof(GeneratorView), typeDiscriminator: EmbeddedViewTypeDiscriminators.GeneratorView)]
-[JsonDerivedType(typeof(ListView), typeDiscriminator: EmbeddedViewTypeDiscriminators.ListView)]
-[JsonDerivedType(typeof(LabelerView), typeDiscriminator: EmbeddedViewTypeDiscriminators.LabelerView)]
-[JsonDerivedType(typeof(StarterPackViewBasic), typeDiscriminator: EmbeddedViewTypeDiscriminators.StarterPackViewBasic)]
 [JsonDerivedType(typeof(EmbeddedView), typeDiscriminator: EmbeddedViewTypeDiscriminators.EmbedView)]
 public record View : Bluesky.View
 {